add hashes

+++ src/type.h
@@ -1,8 +1,8 @@
 #ifndef TYPE_H
 #define TYPE_H
 
-typedef unsigned long long hash_val;
-typedef const char        *hash_key;
-typedef       char        *hash_t  ;
+typedef unsigned long hash_val;
+typedef const char   *hash_key;
+typedef       char   *hash_t  ;
 
 #endif //TYPE_H
